requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
audio-recorder-streamlit==0.0.10
av==12.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chromadb==0.3.29
click==8.1.8
clickhouse-connect==0.8.9
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
duckdb==1.1.3
entrypoints==0.4
fastapi==0.85.1
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.6.0
google_search_results==2.4.2
greenlet==3.1.1
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.2.10
lxml==5.3.1
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.21.1
numpy==1.26.4
onnxruntime==1.20.1
openai==1.59.6
orjson==3.10.14
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
posthog==3.7.4
propcache==0.2.1
protobuf==5.29.3
pulsar-client==3.5.0
pyarrow==18.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
pylibsrtp==0.10.0
PyMuPDF==1.25.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
six==1.17.0
slack_sdk==3.34.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.20.4
streamlit==1.41.1
streamlit-audiorecorder==0.0.6
streamlit-webrtc==0.47.9
SudachiDict-full==20241021
SudachiPy==0.6.10
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.32.1
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.0
websockets==14.1
yarl==1.18.3
zstandard==0.23.0
